implement edit logic. added styles for edit button. refactoring

+++ todo.js
@@ -8,32 +8,36 @@ let uniqueId = 0;
 let allTodos = [];
 
 const init = () => {
-  todoForm.addEventListener("submit", handleAddSubmit);
+  todoForm.addEventListener("submit", handleAddTodoSubmit);
   todoSelect.addEventListener("change", filterTodos);
   todoListUl.addEventListener("click", handleTodoListClick);
+  todoListUl.addEventListener("input", handleTodoListInput);
 }
 
-const handleAddSubmit = (e) => {
+const handleAddTodoSubmit = (e) => {
   e.preventDefault();
   addTodo();
 }
 
 const filterTodos = (e) => {
   const filterValue = e.target.value;
-  renderTodoList(allTodos.filter(todo => filterValue === 'all' || todo.status === filterValue));
+  const filteredTodos = allTodos.filter(todo => filterValue === 'all' || todo.status === filterValue);
+  renderTodoList(filteredTodos);
 }
 
 const handleTodoListClick = (e) => {
-  const { target } = e;
+  const todoIndex = parseInt(e.target.dataset.id);
+  
+  if (e.target.matches(".delete-button")) deleteTodo(todoIndex);
+  if (e.target.matches(".checkbox")) toggleCheckbox(todoIndex, e.target.checked);
+  if (e.target.matches(".edit-button")) enableEditMode(todoIndex);
+  if (e.target.matches(".edit-save-button")) saveEdit(todoIndex);
+}
 
-  if (target.classList.contains("delete-button")) {
-    const todoIndex = parseInt(target.dataset.id);
-    deleteTodo(todoIndex);
-  }
-
-  if (target.classList.contains("checkbox")) {
-    const todoIndex = parseInt(target.dataset.id);
-    toggleCheckbox(target, todoIndex);
+const handleTodoListInput = (e) => {
+  if (e.target.classList.contains("edit-input")) {
+    const todoIndex = parseInt(e.target.dataset.id);
+    allTodos = allTodos.map(todo => todo.id === todoIndex ? { ...todo, text: e.target.value } : todo);
   }
 }
 
@@ -42,64 +46,89 @@ const deleteTodo = (todoIndex) => {
   renderTodoList(allTodos);
 }
 
-const toggleCheckbox = (checkbox, todoIndex) => {
+const toggleCheckbox = (todoIndex, isChecked) => {
   const todo = allTodos.find(todo => todo.id === todoIndex);
-  todo.status = checkbox.checked ? 'completed' : 'uncompleted';
+  todo.status = isChecked ? 'completed' : 'uncompleted';
+  renderTodoList(allTodos);
+};
+
+const enableEditMode = (todoIndex) => {
+  const todo = allTodos.find(todo => todo.id === todoIndex);
+  todo.isEditing = true;
+  renderTodoList(allTodos);
+}
+
+const saveEdit = (todoIndex) => {
+  const todo = allTodos.find(todo => todo.id === todoIndex);
+  todo.isEditing = false;
   renderTodoList(allTodos);
 }
 
 const addTodo = () => {
   const todoText = todoInput.value.trim();
+  if (!todoText) return alert('Please, enter a non-empty value');
+
   todoInput.value = '';
 
-  if (todoText) {
-    uniqueId += 1;
-    allTodos.push({ id: uniqueId, text: todoText, status: defaultStatus });
-    renderTodoList(allTodos);
-  } else {
-    alert('Please, enter a non-empty value');
-  }
+  uniqueId++;
+  allTodos.push({ id: uniqueId, text: todoText, status: defaultStatus, isEditing: false });
+  todoInput.value = '';
+  renderTodoList(allTodos);
 }
+
+const createButton = (text, className, dataId) => {
+  const button = document.createElement("button");
+  button.textContent = text;
+  button.className = className;
+  button.dataset.id = dataId;
+  return button;
+};
+
+const createInput = (type, className, value, dataId) => {
+  const input = document.createElement("input");
+  input.type = type;
+  input.className = className;
+  input.value = value;
+  input.dataset.id = dataId;
+  return input;
+};
+
+const createTodoItem = (todo) => {
+  const todoLi = document.createElement("li");
+  todoLi.id = `todo-${todo.id}`;
+  todoLi.className = `todo ${todo.isEditing ? 'edit-mode' : ''}`;
+
+  const inputsContainer = document.createElement("div");
+  inputsContainer.className = 'inputs-container';
+
+  if (todo.isEditing) {
+    const editInput = createInput("text", "edit-input", todo.text, todo.id);
+    const saveButton = createButton("Save", "edit-save-button", todo.id);
+    inputsContainer.append(editInput, saveButton);
+    todoLi.append(inputsContainer, createButton("Delete", "delete-button", todo.id));
+  } else {
+    const checkbox = createInput("checkbox", "checkbox", "", todo.id);
+    checkbox.checked = todo.status === 'completed';
+
+    const labelText = document.createElement("label");
+    labelText.className = 'todo-text';
+    labelText.textContent = todo.text;
+    if (todo.status === 'completed') {
+      labelText.style.textDecoration = 'line-through';
+      labelText.style.color = '#005f7a';
+    }
+
+    todoLi.append(checkbox, labelText, createButton("Edit", "edit-button", todo.id), createButton("Delete", "delete-button", todo.id));
+  }
+
+  return todoLi;
+};
 
 const renderTodoList = (todos) => {
   todoListUl.innerHTML = '';
   todos.forEach(todo => {
     todoListUl.appendChild(createTodoItem(todo));
   });
-}
-
-const createTodoItem = (todo) => {
-  const todoLi = document.createElement("li");
-  todoLi.id = `todo-${todo.id}`;
-  todoLi.className = 'todo';
-
-  const checkbox = document.createElement("input");
-  checkbox.type = "checkbox";
-  checkbox.className = 'checkbox';
-  checkbox.dataset.id = todo.id;
-  checkbox.checked = todo.status === 'completed';
-
-  const labelText = document.createElement("label");
-  labelText.className = 'todo-text';
-  labelText.textContent = todo.text;
-
-  if (todo.status === 'completed') {
-    labelText.style.textDecoration = 'line-through';
-    labelText.style.color = '#005f7a';
-  }
-
-  const deleteButton = document.createElement("button");
-  deleteButton.className = 'delete-button';
-  deleteButton.textContent = 'Delete';
-  deleteButton.dataset.id = todo.id;
-
-  const editButton = document.createElement("button");
-  editButton.className = 'edit-button';
-  editButton.textContent = 'Edit';
-
-  todoLi.append(checkbox, labelText, editButton, deleteButton);
-
-  return todoLi;
 }
 
 init();
